test: Update current SummaryList tests
